move consoleService into app/models folder
